Version 2.0
Allows launched application to be in a sub folder.
Allows launcher background to be updated, and prioritizes downloading the background. Immediately shows the background as soon as it's downloaded.
Does not crash if you try updating the updater executable.
Settings.json can be in any subdirectory and the updater will find it.
If you set the updaters icon via resource hacker, that icon will also show in the update window.